# openapi.yaml
openapi: "3.1.0"
info:
  title: "Aegis API"
  version: "1.0.0"
  description: "Orchestrate Claude Code sessions via REST API. Aegis wraps Claude Code in ACP sessions\nand exposes everything through a unified API.\n\n**Base URL:** `http://localhost:9100`\n\n**Authentication:** All endpoints except `/v1/health` and `/health` require a bearer token.\nPass `Authorization: Bearer <your-token>` header, or set `AEGIS_AUTH_TOKEN` env var.\n\n**Session lifecycle:** `working` → `idle` → `permission_prompt` → `asking` → `stalled`\n\n**API versioning:** All `/v1/` responses include `X-Aegis-API-Version: 1`.\nLegacy unversioned paths (e.g. `/health`, `/sessions`) emit `Deprecation: true`,\n`Sunset`, and `X-API-Deprecated` headers. See `docs/api-versioning.md` for the\nfull versioning policy.\n"
servers:
  - url: "http://localhost:9100"
    description: "Local Aegis server"
security:
  - BearerAuth: []
paths:
...
  /v1/sessions:
    get:
      operationId: "listSessions"
      summary: "List all sessions"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/limit"

        - $ref: "#/components/parameters/offset"

        - name: "status"
          in: "query"
          schema:
            $ref: "#/components/schemas/SessionStatusFilter"

          description: "Filter by session UI state"


      responses:
        "200":
          description: "List of sessions"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  sessions:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/SessionInfo"


                  total:
                    type: "integer"






        "401":
          $ref: "#/components/responses/Unauthorized"



    post:
      operationId: "createSession"
      summary: "Create or reuse a session"
      tags:
        - "Sessions"

      description: "Creates a new Claude Code session or reuses an existing idle session for the same `workDir`.\nReturns `201` for new sessions, `200` for reused ones.\n"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"




      responses:
        "200":
          description: "Existing idle session reused"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionReused"




        "201":
          description: "New session created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreated"




        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "409":
          description: "Session creation conflict"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  sessionId:
                    type: "string"
...
  "/v1/sessions/{id}":
    get:
      operationId: "getSession"
      summary: "Get session details"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "200":
          description: "Session details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"




        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"



    delete:
      operationId: "killSession"
      summary: "Kill a session"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "204":
          description: "Session killed"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
  "/v1/sessions/{id}/send":
    post:
      operationId: "sendMessage"
      summary: "Send a message to a session"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                text:
                  type: "string"
                  description: "Message text to send to Claude Code"
                  maxLength: 100000


              required:
                - "text"





      responses:
        "202":
          description: "Message delivered"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  delivered:
                    type: "boolean"

                  attempts:
                    type: "integer"






        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"

        "409":
          description: "Session not in a state that accepts messages"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  status:
                    type: "string"
  "/v1/sessions/{id}/read":
    get:
      operationId: "readMessages"
      summary: "Read session transcript"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"

        - name: "offset"
          in: "query"
          schema:
            type: "integer"
            default: 0

          description: "Byte offset to start reading from"

        - name: "limit"
          in: "query"
          schema:
            type: "integer"
            default: 50
            minimum: 1
            maximum: 200



      responses:
        "200":
          description: "Parsed transcript entries"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  messages:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/TranscriptEntry"


                  hasMore:
                    type: "boolean"

                  nextOffset:
                    type: "integer"






        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
...
  "/v1/sessions/{id}/children":
    get:
      operationId: "getChildren"
      summary: "List child sessions spawned from this session"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "200":
          description: "Child sessions"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  children:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/SessionInfo"







        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
  "/v1/sessions/{id}/spawn":
    post:
      operationId: "spawnChild"
      summary: "Spawn a child session from a running session"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                workDir:
                  type: "string"

                prompt:
                  type: "string"

                permissionMode:
                  type: "string"
                  enum:
                    - "default"
                    - "bypassPermissions"
                    - "plan"
                    - "acceptEdits"
                    - "dontAsk"
                    - "auto"







      responses:
        "201":
          description: "Child session created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreated"




        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
  "/v1/sessions/{id}/fork":
    post:
      operationId: "forkSession"
      summary: "Fork a session — clone context into a new session"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "201":
          description: "Forked session created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreated"




        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
  "/v1/sessions/{id}/permissions":
    get:
      operationId: "getPermissionPolicy"
      summary: "Get session's permission policy"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "200":
          description: "Permission policy"
          content:
            application/json:
              schema:
                type: "object"




        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"



    put:
      operationId: "updatePermissionPolicy"
      summary: "Update session's permission policy"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"




      responses:
        "200":
          description: "Policy updated"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
  "/v1/sessions/{id}/permission-profile":
    get:
      operationId: "getPermissionProfile"
      summary: "Get session's permission profile"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "200":
          description: "Permission profile"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  profile:
                    type: "string"
                    enum:
                      - "allow"
                      - "deny"
                      - "ask"







        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"



    put:
      operationId: "updatePermissionProfile"
      summary: "Update session's permission profile"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                profile:
                  type: "string"
                  enum:
                    - "allow"
                    - "deny"
                    - "ask"



              required:
                - "profile"





      responses:
        "200":
          description: "Profile updated"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
...
  "/v1/sessions/{id}/interrupt":
    post:
      operationId: "interruptSession"
      summary: "Send Ctrl+C to interrupt Claude Code"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "202":
          description: "Interrupt sent"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
  "/v1/sessions/{id}/escape":
    post:
      operationId: "sendEscape"
      summary: "Send Escape key"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "202":
          description: "Escape sent"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
...
  "/v1/sessions/{id}/command":
    post:
      operationId: "sendCommand"
      summary: "Send a shell command directly"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                command:
                  type: "string"
                  description: "Shell command to execute"

                workingDirectory:
                  type: "string"


              required:
                - "command"





      responses:
        "202":
          description: "Command sent"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
...
  /sessions:
    get:
      operationId: "listSessionsAlias"
      summary: "List all sessions (alias, deprecated)"
      deprecated: true
      tags:
        - "Sessions"

      responses:
        "200":
          description: "List of sessions"
          headers:
            Deprecation:
              description: "Indicates this endpoint is deprecated (RFC 8594)"
              schema:
                type: "string"
                example: "true"


            Sunset:
              description: "Date after which this endpoint may be removed"
              schema:
                type: "string"
                format: "date-time"
                example: "Thu, 01 Jan 2027 00:00:00 GMT"


            X-API-Deprecated:
              description: "Human-readable replacement path"
              schema:
                type: "string"
                example: "Use /v1/sessions instead"



          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/SessionInfo"





        "401":
          $ref: "#/components/responses/Unauthorized"
...
  "/v1/sessions/{id}/answer":
    post:
      summary: "Answer pending question"
      operationId: "answerSessionQuestion"
      tags:
        - "Session Actions"

      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Session UUID"
          schema:
            type: "string"
            format: "uuid"
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"



      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                questionId:
                  type: "string"
                  minLength: 1

                answer:
                  type: "string"


              required:
                - "questionId"
                - "answer"

              additionalProperties: false




      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"


                required:
                  - "ok"

                additionalProperties: false




        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"

                  details:
                    type: "array"
                    items: {}


                required:
                  - "error"

                additionalProperties: false




        "409":
          description: "No pending question matching this questionId"
...
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "API key"
  parameters:
    sessionId:
      name: "id"
      in: "path"
      required: true
      schema:
        type: "string"

      description: "Session ID"
    limit:
      name: "limit"
      in: "query"
      schema:
        type: "integer"
        default: 50
        minimum: 1
        maximum: 200
    offset:
      name: "offset"
      in: "query"
      schema:
        type: "integer"
        default: 0
  responses:
    BadRequest:
      description: "Invalid request body or parameters"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
    Unauthorized:
      description: "Missing or invalid authentication token"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Unauthorized"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Not found"
  schemas:
...
    CreateSessionRequest:
      type: "object"
      properties:
        name:
          type: "string"
          description: "Human-readable session name"
          example: "feature-auth"

        workDir:
          type: "string"
          description: "Working directory for Claude Code"
          example: "/home/user/my-project"

        prompt:
          type: "string"
          description: "Initial prompt to send to Claude Code"

        permissionMode:
          type: "string"
          enum:
            - "default"
            - "bypassPermissions"
            - "plan"
            - "acceptEdits"
            - "dontAsk"
            - "auto"

          default: "default"

        autoApprove:
          type: "boolean"
          default: false

        env:
          type: "object"
          additionalProperties:
            type: "string"

          description: "Environment variables to inject into the session"

        claudeCommand:
          type: "string"
          maxLength: 10000
          description: "Custom Claude Code command/flags"

        maxSessionAgeMs:
          type: "integer"
          description: "Maximum session lifetime in milliseconds"

        stallThresholdMs:
          type: "integer"
          description: "Stall detection threshold override"


      required:
        - "workDir"
    SessionCreated:
      type: "object"
      properties:
        id:
          type: "string"

        name:
          type: "string"

        status:
          type: "string"
          example: "working"

        workDir:
          type: "string"

        reused:
          type: "boolean"
          example: false

        promptDelivery:
          type: "object"
          properties:
            delivered:
              type: "boolean"

            attempts:
              type: "integer"
    SessionReused:
      allOf:
        - $ref: "#/components/schemas/SessionCreated"

        - type: "object"
          properties:
            reused:
              type: "boolean"
              example: true
    SessionInfo:
      type: "object"
      properties:
        id:
          type: "string"

        windowId:
          type: "string"

        windowName:
          type: "string"

        workDir:
          type: "string"

        claudeSessionId:
          type: "string"

        jsonlPath:
          type: "string"

        byteOffset:
          type: "integer"

        monitorOffset:
          type: "integer"

        status:
          type: "string"
          enum:
            - "idle"
            - "working"
            - "compacting"
            - "context_warning"
            - "waiting_for_input"
            - "permission_prompt"
            - "plan_mode"
            - "ask_question"
            - "bash_approval"
            - "settings"
            - "error"
            - "rate_limit"
            - "unknown"


        createdAt:
          type: "integer"

        lastActivity:
          type: "integer"

        stallThresholdMs:
          type: "integer"

        permissionStallMs:
          type: "integer"

        permissionMode:
          type: "string"

        settingsPatched:
          type: "boolean"

        hookSettingsFile:
          type: "string"

        hookSecret:
          type: "string"

        lastHookAt:
          type: "integer"

        activeSubagents:
          type: "array"
          items:
            type: "string"


        permissionPromptAt:
          type: "integer"

        permissionRespondedAt:
          type: "integer"

        lastHookReceivedAt:
          type: "integer"

        lastHookEventAt:
          type: "integer"

        model:
          type: "string"

        lastDeadAt:
          type: "integer"

        ccPid:
          type: "integer"

        parentId:
          type: "string"

        children:
          type: "array"
          items:
            type: "string"


        permissionPolicy:
          type: "array"
          items:
            type: "object"


        permissionProfile:
          type: "object"

        prd:
          type: "string"

        ownerKeyId:
          type: "string"
...
    TranscriptEntry:
      type: "object"
      properties:
        id:
          type: "integer"

        role:
          type: "string"
          enum:
            - "user"
            - "assistant"
            - "system"


        content:
          type: "string"

        timestamp:
          type: "integer"

        toolCalls:
          type: "array"
          items:
            type: "object"
...
    SessionStatusFilter:
      type: "string"
      enum:
        - "all"
        - "idle"
        - "working"
        - "compacting"
        - "context_warning"
        - "waiting_for_input"
        - "permission_prompt"
        - "plan_mode"
        - "ask_question"
        - "bash_approval"
        - "settings"
        - "error"
        - "rate_limit"
        - "unknown"