# src/routes/session-actions.ts
/**
 * routes/session-actions.ts — Session operations: send, read, answer,
 * escape, interrupt, kill, command, children, spawn, fork, permissions.
 */

import type { FastifyInstance, FastifyRequest, FastifyReply } from 'fastify';
import { sendMessageSchema, commandSchema, permissionRuleSchema, permissionProfileSchema, type PermissionProfile } from '../validation.js';
import type { PermissionPolicy } from '../validation.js';
import { registerPermissionRoutes } from '../permission-routes.js';
import { cleanupTerminatedSessionState } from '../session-cleanup.js';
import {
  type RouteContext,
  makePayload,
  registerWithLegacy,
  requirePermission,
  resolveRequestAuditActor,
  getRequestRole,
  requestHasPermission,
  withOwnership,
  withSessionOwnership,
} from './context.js';

export function registerSessionActionRoutes(app: FastifyInstance, ctx: RouteContext): void {
  const {
    sessions, auth, quotas, config, metrics, monitor, eventBus, channels,
    toolRegistry, getAuditLogger, validateWorkDir,
  } = ctx;

  // Send message (with delivery verification — Issue #1)
  // Issue #2461: Extracted to variable so /input alias reuses the same handler.
  const sendHandler = withSessionOwnership(ctx, async (req, reply, session) => {
    if (!requirePermission(auth, req, reply, 'send')) return;
    const parsed = sendMessageSchema.safeParse(req.body);
    if (!parsed.success) return reply.status(400).send({ error: 'Invalid request body', details: parsed.error.issues });
    const { text } = parsed.data;
    const sessionId = session.id;

    // Issue #1953: Per-key token/spend quota enforcement at send time.
    const keyId = req.authKeyId;
    const apiKey = keyId && keyId !== 'master' ? auth.getKey(keyId) : null;
    if (apiKey) {
      const ownedSessions = sessions.listSessions().filter(s => s.ownerKeyId === keyId);
      const quotaResult = quotas.checkSendQuota(apiKey, ownedSessions.length);
      if (!quotaResult.allowed) {
        return reply.status(429).send({
          error: 'QUOTA_EXCEEDED',
          message: quotaResult.message,
          quota: quotaResult.reason,
          usage: quotaResult.usage,
        });
      }
    }
    try {
      const result = await sessions.sendMessage(sessionId, text);
      // Issue #1809: Re-fetch stall info AFTER delivery to avoid false-positive.
      // Previously we called getStallInfo BEFORE send, capturing a stale state
      // (session was temporarily quiet but became active after message delivery).
      const currentStallInfo = monitor.getStallInfo(sessionId);
      await channels.message({
        event: 'message.user',
        timestamp: new Date().toISOString(),
        session: { id: sessionId, name: '', workDir: '' },
        detail: text,
      });
      const response: Record<string, unknown> = { ok: true, delivered: result.delivered, attempts: result.attempts };
      if (!result.delivered) {
        response.reason = result.error ?? 'no_active_transport';
      }
      if (currentStallInfo.stalled) response.stall = currentStallInfo;
      return response;
    } catch (e: unknown) {
      return reply.status(404).send({ error: e instanceof Error ? e.message : String(e) });
    }
  }, 'send');

  registerWithLegacy(app, 'post', '/v1/sessions/:id/send', sendHandler);

  // Issue #2461: /input alias for /send
  registerWithLegacy(app, 'post', '/v1/sessions/:id/input', sendHandler);

  // Issue #702: GET children sessions
  registerWithLegacy(app, 'get', '/v1/sessions/:id/children', withOwnership(sessions, async (_req, _reply, session) => {
    const children = (session.children ?? []).map(id => {
      const child = sessions.getSession(id);
      if (!child) return null;
      return { id: child.id, displayName: child.displayName, status: child.status, createdAt: child.createdAt };
    }).filter(Boolean);
    return { children };
  }));

  // Issue #702: Spawn child session
  interface SpawnBody { name?: string; prompt?: string; workDir?: string; permissionMode?: string; }
  registerWithLegacy(app, 'post', '/v1/sessions/:id/spawn', withOwnership(sessions, async (req, reply, parent) => {
    if (!requirePermission(auth, req, reply, 'create')) return;
    const { name, prompt, workDir, permissionMode } = (req.body as SpawnBody | undefined) ?? {};
    const childName = name ?? `${parent.displayName ?? 'session'}-child`;
    const requestedWorkDir = workDir ?? parent.workDir;
    const safeChildWorkDir = await validateWorkDir(requestedWorkDir);
    if (typeof safeChildWorkDir === 'object') {
      return reply.status(400).send({ error: `Invalid workDir: ${safeChildWorkDir.error}`, code: safeChildWorkDir.code });
    }
    const childPermMode = permissionMode ?? parent.permissionMode ?? 'default';
    const childSession = await sessions.createSession({ workDir: safeChildWorkDir, name: childName, parentId: parent.id, permissionMode: childPermMode, ownerKeyId: req.authKeyId });
    let promptDelivery: { delivered: boolean; attempts: number } | undefined;
    if (prompt) { promptDelivery = await sessions.sendInitialPrompt(childSession.id, prompt); }
    return reply.status(201).send({ ...childSession, promptDelivery });
  }));

  // Issue #468: Fork session
  interface ForkBody { name?: string; prompt?: string; clearPanes?: boolean; }
  registerWithLegacy(app, 'post', '/v1/sessions/:id/fork', withOwnership(sessions, async (req, reply, parent) => {
    if (!requirePermission(auth, req, reply, 'create')) return;
    const { name, prompt } = (req.body as ForkBody | undefined) ?? {};
    const forkName = name ?? `${parent.displayName ?? 'session'}-fork`;
    const forkedSession = await sessions.createSession({
      workDir: parent.workDir,
      name: forkName,
      permissionMode: parent.permissionMode,
      ownerKeyId: req.authKeyId,
    });
    let promptDelivery: { delivered: boolean; attempts: number } | undefined;
    if (prompt) { promptDelivery = await sessions.sendInitialPrompt(forkedSession.id, prompt); }
    await channels.sessionCreated({
      event: 'session.created',
      timestamp: new Date().toISOString(),
      session: { id: forkedSession.id, name: forkedSession.displayName, workDir: parent.workDir },
      detail: `Session forked from ${parent.id}`,
    });
    return reply.status(201).send({ ...forkedSession, forkedFrom: parent.id, promptDelivery });
  }));

  // Issue #700: Permission policy endpoints
  registerWithLegacy(app, 'get', '/v1/sessions/:id/permissions', withOwnership(sessions, async (_req, _reply, session) => {
    return { permissionPolicy: session.permissionPolicy ?? [] };
  }));
  registerWithLegacy(app, 'put', '/v1/sessions/:id/permissions', withOwnership(sessions, async (req, reply, session) => {
    const policy = (req.body as PermissionPolicy | undefined) ?? [];
    const result = permissionRuleSchema.array().safeParse(policy);
    if (!result.success) return reply.status(400).send({ error: 'Invalid permission policy', details: result.error.issues });
    session.permissionPolicy = policy;
    await sessions.save();
    return { permissionPolicy: policy };
  }));

  registerWithLegacy(app, 'get', '/v1/sessions/:id/permission-profile', withOwnership(sessions, async (_req, _reply, session) => {
    return { permissionProfile: session.permissionProfile ?? null };
  }));
  registerWithLegacy(app, 'put', '/v1/sessions/:id/permission-profile', withOwnership(sessions, async (req, reply, session) => {
    const parsed = permissionProfileSchema.safeParse((req.body as PermissionProfile | undefined) ?? {});
    if (!parsed.success) return reply.status(400).send({ error: 'Invalid permission profile', details: parsed.error.issues });
    session.permissionProfile = parsed.data;
    await sessions.save();
    return { permissionProfile: parsed.data };
  }));

  // Read messages
  // Issue #2539: Use readMessagesFromSession() to avoid the TOCTOU race between
  // the ownership check (which already resolved the session) and readMessages()
  // doing a second this.state.sessions[id] lookup. Also map non-404 errors to
  // 500 so transient runtime errors don't masquerade as missing sessions.
  registerWithLegacy(app, 'get', '/v1/sessions/:id/read', withOwnership(sessions, async (_req, reply, session) => {
    try {
      return await sessions.readMessagesFromSession(session);
    } catch (e: unknown) {
      return reply.status(500).send({ error: e instanceof Error ? e.message : String(e) });
    }
  }));

  // Register approve/reject permission routes
  registerPermissionRoutes(
    app,
    {
      approve: async (id: string) => sessions.approve(id),
      reject: async (id: string) => sessions.reject(id),
      getLatencyMetrics: (id: string) => sessions.getLatencyMetrics(id),
      getSession: (id: string) => sessions.getSession(id),
    },
    {
      recordPermissionResponse: (id: string, latencyMs: number) => metrics.recordPermissionResponse(id, latencyMs),
    },
    null,
    {
      getAuditLogger: () => getAuditLogger() ?? null,
      hasPermission: (_keyId, permission, req) => requestHasPermission(auth, req, permission),
      resolveRole: (_keyId, req) => getRequestRole(auth, req),
      config,
    },
  );

  // Issue #336: Answer pending AskUserQuestion
  registerWithLegacy(app, 'post', '/v1/sessions/:id/answer', withOwnership(sessions, async (req: FastifyRequest, reply: FastifyReply, session) => {
    if (!requirePermission(auth, req, reply, 'send')) return;
    const { questionId, answer } = (req.body as { questionId?: string; answer?: string } | undefined) || {};
    if (!questionId || answer === undefined || answer === null) {
      return reply.status(400).send({ error: 'questionId and answer are required' });
    }
    const resolved = sessions.submitAnswer(session.id, questionId, answer);
    if (!resolved) {
      return reply.status(409).send({ error: 'No pending question matching this questionId' });
    }
    return { ok: true };
  }));

  // Escape
  registerWithLegacy(app, 'post', '/v1/sessions/:id/escape', withSessionOwnership(ctx, async (req, reply, session) => {
    if (!requirePermission(auth, req, reply, 'send')) return;
    try {
      await sessions.escape(session.id);
      return { ok: true };
    } catch (e: unknown) {
      return reply.status(404).send({ error: e instanceof Error ? e.message : String(e) });
    }
  }, 'send'));

  // Interrupt (Ctrl+C)
  registerWithLegacy(app, 'post', '/v1/sessions/:id/interrupt', withSessionOwnership(ctx, async (req, reply, session) => {
    if (!requirePermission(auth, req, reply, 'send')) return;
    try {
      await sessions.interrupt(session.id);
      return { ok: true };
    } catch (e: unknown) {
      return reply.status(404).send({ error: e instanceof Error ? e.message : String(e) });
    }
  }, 'send'));

  // Kill session
  // Issue #2461: Extracted to variable so POST aliases reuse the same handler.
  const killHandler = withSessionOwnership(ctx, async (req, reply, session) => {
    if (!requirePermission(auth, req, reply, 'kill')) return;
    try {
      await sessions.killSession(session.id);
      // Issue #2067: record session as failed before cleanup
      metrics.sessionFailed(session.id);
      eventBus.emitEnded(session.id, 'killed');
      const auditLogger = getAuditLogger();
      if (auditLogger) void auditLogger.log(resolveRequestAuditActor(auth, req, 'system'), 'session.kill', `Session killed: ${session.id} (permission=${req.matchedPermission ?? 'kill'})`, session.id, req.tenantId);
      await channels.sessionEnded(makePayload(sessions, 'session.ended', session.id, 'killed'));
      cleanupTerminatedSessionState(session.id, { monitor, metrics, toolRegistry });
      return { ok: true };
    } catch (e: unknown) {
      return reply.status(404).send({ error: e instanceof Error ? e.message : String(e) });
    }
  }, 'kill');

  registerWithLegacy(app, 'delete', '/v1/sessions/:id', killHandler);

  // Issue #2461: POST aliases for kill
  for (const alias of ['/v1/sessions/:id/kill', '/v1/sessions/:id/terminate', '/v1/sessions/:id/stop'] as const) {
    registerWithLegacy(app, 'post', alias, killHandler);
  }

  // Slash command
  registerWithLegacy(app, 'post', '/v1/sessions/:id/command', withSessionOwnership(ctx, async (req, reply, session) => {
    if (!requirePermission(auth, req, reply, 'send')) return;
    const parsed = commandSchema.safeParse(req.body);
    if (!parsed.success) return reply.status(400).send({ error: 'Invalid request body', details: parsed.error.issues });
    const { command } = parsed.data;
    try {
      const cmd = command.startsWith('/') ? command : `/${command}`;
      await sessions.sendMessage(session.id, cmd);
      return { ok: true };
    } catch (e: unknown) {
      return reply.status(404).send({ error: e instanceof Error ? e.message : String(e) });
    }
  }, 'send'));

}